Rename RpcApiGrpc to GrpcClient and refactor Resolver
